requirements.txt
pandas==2.2.2
attrs==20.3.0
bcrypt==3.2.0
numpy==1.24.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
configparser==5.0.2
cryptography==3.4.7
distro==1.5.0
docker==5.0.0
docker-compose==1.29.0
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==7.12.0
idna==2.10
importlib-metadata==3.10.0
itsdangerous==2.0.0
Jinja2==3.0
jsonschema==3.2.0
MarkupSafe==2.1.5
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.17.0
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
texttable==1.6.3
typing-extensions==4.12.2
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==2.2.0
zipp==3.4.1
transformers==4.34.0
sentence-transformers==3.0.1
fastapi==0.65.2
uvicorn==0.30.0
